Adding a shell script to work around mpiexec closing stdin.
The problem that I have discovered is that mpiexec closes stdin (file number 0).  Then it runs RAVEN.  Some library that raven calls opens "/" and assigns that to file number 0.  Then when python runs, it see that stdin (file number 0) is a directory, and errors out.

This just has mpiexec call a shell script that uses /dev/zero as stdin,
which python is fine with.

r21367