Adding the ability to specifiy the random seed in the step.
This adds to the existing ability to specify the seed in the sampler.
Now, if the sampler does not specify an initial seed, the random number
generator is not changed.  Also, now, the boost random generator is always
used, instead of sometimes the numpy one being used.

In addition, the sampler now can take two keywords: "reseed" and "continue"
reseed re-randomizes the random number generator (in the way the sampler used
to always) and continue does not touch the random number generator.

r24896