Enableing initial support of grid and latin hyper cube for raven applications in the framework.
This does the modification of the distributions so the the grid works.
Note that this does not yet check that the distributions in the .i file (if
they exist) are consistent with the ones in the .xml file.
Note that varibles and distributions are not added to the
csv file, unlike the MOOSE apps version.

r25697