Switching ControlLogicBranchingInfo to a FileOutput.
This changes ControlLogicBranchingInfo into a subclass of FileOutput,
so that the files names can be correct.  This also changes RAVENInterface.py
to add the ControlLogicBranchingInfo as an output.  This allows
dynamic event trees to work again, so the tests are no longer skipped.

This still needs work, since it should probably save the branch info
when there is branch info, instead of when RavenExecutioner explicitly
calls into ControlLogicBranchingInfo.

r28044