Put finding raven libs directory in if block.
This puts the check for RAVEN_LIBS_DIR inside an if block that first checks
if the upper directory exists.  Otherwise the error:
ls: cannot access .../raven_libs/pylibs/*/python*/site-packages: No such file or directory
is printed out, which can confuse users trying to debug other issues.
Finding out which libraries are used is better done with the
./run_tests --library-report
command.
Closes #385