Move all functions that use numpy to mathUtils.
Basically, find_crow is used to find crow, but if things use numpy in
utils.py, find_crow cannot be used if numpy is not available.
This is a problem since --library-report uses find_crow, and
therefore will fail if numpy is not available.
This moves the functions to mathUtils which already is importing scipy.
Closes #486