Add files via upload
Corrections made: 
in mrtau standalone file: two variables were added for more complete testing 
in phisics file: some of the variables were transparent (modified by raven, but non existent in the input, so they were not replaced). That was the case for the DENSITY variable and the FY variables. A FY variable in the fast range was also added.