Plugins fixes (#1115)
* Closes #1111 adds warning for no test sets run

Adds a check for whether any test sets (e.g. RAVEN, Cashflow) were run.
If not, then a warning is issued. This could potentially be changed
to return a fail value instead of a warning.

* Closes #1112 set a temp workdir for plugin install

Adds a temporary working directory change to "raven" root dir for the
duration of the plugin install script, to assure it works from an
arbitrary location.