Updated Validation.py
Updated to request list type data from evaluation. The change enables to create a csv per sample for HistorySet type.
This will require for every validation code to return a list of dictionaries. Each dictionary is for the given sample.
If the output of the validation code only returns a dictionary for all samples, add only one dictionary to the list.