Update ROM.py
When the LearningGate.py was merged with ROM.py, the derivative function in the LearningGate.py was not added. The derivative function in the ROM.py was originally using the derivative function from the LearningGate,py which was deprecated. The derivative function in ROM.py now directly uses the mathUtils as the LearningGate.py did.

Since the self.supervisedEngine is now an artifact of past ROM.py versions, it has been deleted in this commit.