Parallel improvements (#1825)
* If ray instatiated outside, use it.

Basically, before, ray was only used if nodes was setup,
such as with MPI mode.

* No longer automatically adds mpi mode to inner ravens.

* Make it a bigger test, and switch to internalParallel

* Adding useful debugging information.

* Wait for servers to finish starting.

Otherwise they become zombies.

* Only print out changed part.

* Adding debugging info, and force status to disk.

* Setting port to 0 so ray chooses an available port.

This adds port as a parameter to the starting ray function,
and sends in 0.  This tells ray to choose an available port,
instead of erroring if port 6379 is not available.

* Find correct ray start

`ray start sometimes appears, so add a space before it.

* Don't start JobHandler if running remotely.

If we are going to be running remotely, then we should
not start job handler. Otherwise, we have to start ray or threads, only
to seconds later shutdown them down.