Removing assumption that python is available when running --install
Moved installing the Example plugin till after the environment is created
since install_plugins.py needs the python command to exist.

Added logic to the conda install path to find a python command that is
in the PATH, by checking PYTHON_COMMAND, python and python3, and using
the first that is available.