Slurm addition (#2396)
* Starting work on Slurm mode.

* Adding first slurm test.

* Ignore slurm output files.

* Adding script to run slurm tests.

* Adding test_mpislurm_long

* Updating RavenRunsRaven tests to check on slurm.

* Adding back runQSUB to some tests it was missing from.

* Changing MPISimulationMode to PBSSimulationMode.py

This is to reflect that it is pbs specific.

* Updating the manual to include slurm.

Also changing mode name to pbs from mpi, since both use mpi.

* Removing unused queueingSoftware node.

This is not used by any standard mode, so commenting out documentation.
The code is left in Simulation.py incase there are any CustomModes outside of
RAVEN that use it.

* Support when nodenames are not fqdn.

* Adding more memory to Ray run.

* Various parallel fixes.

Remove extra space that caused problems with command line.
Call ray.shutdown before closing servers so it can succeed.
Run ray stop's in parallel to speed up shutdown.

* Adding --longest-jobs for compatibility.

* Switching to using new pbs mode name.

* Adding printing out time that raven starts.

* Check for existance of module files before sourcing.

* Updating ray version.

* Fixing so ray works on one node.

* Fixing bug where verbosity got reset if no verbosity in file.

* Fixing review comments.

* Support using InputSpecs and convert SlurmSimulationMode to use.

* Making input parsing consistent with manual.